refactor: item 디버그 수정

+++ OpenMarket/OpenMarket/Model/JSONModel/ProductDetail.swift
@@ -13,7 +13,7 @@ struct ProductDetail: Decodable {
     let currency: Currency
     let price: Double
     let bargainPrice: Double
-    let discountedPrice: Int
+    let discountedPrice: Double
     let stock: Int
     let createdAt: String
     let issuedAt: String
